+++ __init__.py
@@ -1,2 +1,2 @@
 """Ohmatic: AI-Powered Circuit Schematic Generator"""
-__version__ = "0.0.1"
+__version__ = "0.0.2"